# Backend/app/configs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "P0136. GIS - Prototype Development",
    "description": "- Use opensource\n - Using AWS\n - First demo Aug - 01\n - Release POC end of September\n - Start date: 07/11/2019\n - End date: 09/20/2019"
  },
  "host": "localhost:8080",
  "basePath": "/api/v1",
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
    "/users": {
      "get": {
        "tags": ["Users"],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Status of user that we want to find(active or inactive) ex: status=active.",
            "type": "string"
          },
          {
            "name": "sort_field",
            "in": "query",
            "description": "Field of user that we want to sort ex: sort_field=username.",
            "type": "string"
          },
          {
            "name": "sort_type",
            "in": "query",
            "description": "Type of Field that we want to sort(asc or desc) ex: sort_type=asc.",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number, starting with 1. ex: page=1.",
            "type": "number"
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "Requested response Page size. Minimum, maximum and default is defined by Provider. Minimum can never be lower than 1 ex: pageSize=10.",
            "type": "number"
          },
          {
            "name": "keyword",
            "in": "query",
            "description": "Search user by username include keyword.",
            "type": "string"
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "Succesfully retrieved resource.",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "post": {
        "summary": "Create new user in system",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to create",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "New user is created",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/users/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        },
        {
          "name": "Authorization",
          "in": "header",
          "description": "An authorization header .",
          "type": "string",
          "required": true
        }
      ],
      "get": {
        "tags": ["Users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "summary": "Get user with given ID",
        "responses": {
          "200": {
            "description": "User is found",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user with given ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User is deleted",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      },
      "put": {
        "summary": "Update user with give ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User with new values of properties",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/users/change-status": {
      "put": {
        "summary": "Update status multy user with give array ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Status that we want to change",
            "schema": {
              "properties": {
                "status": {
                  "type": "string"
                },
                "userIds": {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                }
              }
            }
          },
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/users/change-status/{userId}": {
      "parameters": [
        {
          "name": "userId",
          "in": "path",
          "required": true,
          "description": "ID of user that we want to find",
          "type": "string"
        },
        {
          "name": "Authorization",
          "in": "header",
          "description": "An authorization header .",
          "type": "string",
          "required": true
        }
      ],
      "put": {
        "summary": "Update status of user with give ID",
        "tags": ["Users"],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "An authorization header .",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User is updated",
            "schema": {
              "$ref": "#/definitions/UserResponse"
            }
          },
          "400": {
            "description": "User is notfound",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
...
    "/auth/login": {
      "post": {
        "summary": "Login into the system with username & password",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "request",
            "in": "body",
            "description": "Status that we want to change",
            "schema": {
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticate success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "access_token": {
                        "type": "string"
                      },
                      "refresh_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/auth/refresh_token": {
      "post": {
        "summary": "Get new token",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "body",
            "schema": {
              "properties": {
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticate success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "properties": {
                      "access_token": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Logout the system",
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "refresh_token",
            "in": "body",
            "schema": {
              "properties": {
                "refresh_token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logout success.",
            "schema": {
              "properties": {
                "status": {
                  "$ref": "#/definitions/Status"
                },
                "results": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  },
                  "example": []
                }
              }
            }
          },
          "400": {
            "description": "Failed to authenticate with token.",
            "schema": {
              "$ref": "#/definitions/NotFound"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "required": ["email", "id", "username", "password"],
      "properties": {
        "id": {
          "type": "number",
          "uniqueItems": true
        },
        "status": {
          "type": "string",
          "enum": ["active", "inactive"]
        },
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "lastname": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "language": {
          "type": "string"
        },
        "roleId": {
          "type": "number"
        },
        "country": {
          "type": "string"
        },
        "last_login": {
          "type": "string"
        }
      }
    },
    "UserResponse": {
      "properties": {
        "status": {
          "$ref": "#/definitions/Status"
        },
        "results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
...
    "Status": {
      "properties": {
        "error": {
          "type": "boolean",
          "example": false
        },
        "code": {
          "type": "number"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "NotFound": {
      "properties": {
        "status": {
          "$ref": "#/definitions/Status"
        },
        "results": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": []
        }
      }
    }
  }
}